Move rollout binding logic into RolloutTimestepPreparationMixin under post_training
Keep the core stage lean, mirroring RolloutDenoisingMixin on DenoisingStage:
scheduler selection, the timesteps/sigmas convention check, and the
once-per-process log flag now live in post_training/rollout_timestep_mixin.py.
The stage keeps only the rollout_scheduler parameter and two call sites.
