+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/timestep_preparation.py
@@ -28,6 +28,9 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages.validators import (
 from sglang.multimodal_gen.runtime.pipelines_core.stages.validators import (
     VerificationResult,
 )
+from sglang.multimodal_gen.runtime.post_training.rollout_timestep_mixin import (
+    RolloutTimestepPreparationMixin,
+)
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 
@@ -45,7 +48,7 @@ class TimestepPreparationFingerprint:
     num_frames: int | None
 
 
-class TimestepPreparationStage(PipelineStage):
+class TimestepPreparationStage(PipelineStage, RolloutTimestepPreparationMixin):
     """
     Stage for preparing timesteps for the diffusion process.
 
@@ -57,10 +60,6 @@ class TimestepPreparationStage(PipelineStage):
     deduplicated_deepcopy_output_fields = ("scheduler",)
     deduplicated_extra_tensor_tree_output_keys = ("mu",)
 
-    # Class-level so the rollout info log prints once per process, not once
-    # per stage instance.
-    _logged_rollout_scheduler_check = False
-
     def __init__(
         self,
         scheduler,
@@ -71,10 +70,7 @@ class TimestepPreparationStage(PipelineStage):
     ) -> None:
         super().__init__()
         self.scheduler = scheduler
-        # Bound per request when rollout=True: the rollout SDE/log-prob path
-        # needs a first-order flow-match Euler scheduler, which not every
-        # pipeline serves (e.g. Wan serves UniPC). None keeps the serving
-        # scheduler for rollout requests.
+        # See RolloutTimestepPreparationMixin.
         self.rollout_scheduler = rollout_scheduler
         self.prepare_extra_set_timesteps_kwargs = list(
             prepare_extra_set_timesteps_kwargs or []
@@ -100,9 +96,11 @@ class TimestepPreparationStage(PipelineStage):
         if batch.scheduler is not None and batch.timesteps is not None:
             return batch
 
-        use_rollout_scheduler = batch.rollout and self.rollout_scheduler is not None
-        template = self.rollout_scheduler if use_rollout_scheduler else self.scheduler
-        scheduler = get_or_create_request_scheduler(batch, template)
+        rollout_template = self._resolve_rollout_scheduler(batch)
+        scheduler = get_or_create_request_scheduler(
+            batch,
+            rollout_template if rollout_template is not None else self.scheduler,
+        )
         device = get_local_torch_device()
         num_inference_steps = batch.num_inference_steps
         timesteps = batch.timesteps
@@ -165,7 +163,7 @@ class TimestepPreparationStage(PipelineStage):
             )
             timesteps = scheduler.timesteps
 
-        if use_rollout_scheduler:
+        if rollout_template is not None:
             self._check_rollout_timesteps(scheduler)
 
         # Update batch with prepared timesteps
@@ -174,32 +172,6 @@ class TimestepPreparationStage(PipelineStage):
         if not batch.is_warmup:
             self.log_debug("timesteps: %s", timesteps)
         return batch
-
-    def _check_rollout_timesteps(self, scheduler) -> None:
-        # The rollout SDE/log-prob math assumes the flow-match Euler
-        # convention timesteps == sigmas[:-1] * num_train_timesteps.
-        sigmas = scheduler.sigmas
-        timesteps = scheduler.timesteps
-        if sigmas is None or timesteps is None or sigmas.numel() < 2:
-            return
-        reconstructed = sigmas[:-1].to(device=timesteps.device) * float(
-            scheduler.config.num_train_timesteps
-        )
-        max_abs_diff = (timesteps.float() - reconstructed.float()).abs().max().item()
-        if max_abs_diff > 1e-3:
-            raise ValueError(
-                f"rollout timestep/sigma mismatch: max_abs_diff={max_abs_diff:.6g}"
-            )
-        if not TimestepPreparationStage._logged_rollout_scheduler_check:
-            logger.info(
-                "RL rollout using %s (timesteps dtype=%s, sigmas dtype=%s, "
-                "max_abs_diff=%.6g)",
-                type(scheduler).__name__,
-                timesteps.dtype,
-                sigmas.dtype,
-                max_abs_diff,
-            )
-            TimestepPreparationStage._logged_rollout_scheduler_check = True
 
     def build_dedup_fingerprint(
         self, batch: Req, server_args: ServerArgs
